Fix some possible bug

+++ pollbot/telegram/commands/poll.py
@@ -18,7 +18,9 @@ from pollbot.models import Poll
 def create_poll(bot, update, session, user):
     """Create a new poll."""
     # The previous unfinished poll will be removed
-    if user.current_poll is not None and not user.current_poll.created:
+    if user.current_poll is not None and  \
+            not user.current_poll.created and \
+            not len(user.current_poll.votes) > 0:
         session.delete(user.current_poll)
 
     poll = Poll(user)
